restructured library
