+++ index.js
@@ -1,4 +1,11 @@
 'use strict';
 
-module.exports.mock = require('./lib/mock/Model');
-module.exports.stub = require('./lib/stub/Model');
+const model = {};
+model.mock = require('./lib/model/model-mock');
+model.stub = require('./lib/model/model-stub');
+module.exports.model = model;
+
+const query = {};
+query.mock = require('./lib/query/query-mock');
+query.stub = require('./lib/query/query-stub');
+module.exports.query = query;
