+++ lib/model/model-mock.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const modelQueryFunctions = require('../model-query-functions');
-const Query = require('./query');
+const modelQueryFunctions = require('./model-query-functions');
+const Query = require('../query/query-mock');
 const deepEqual = require('deep-equal');
 
 module.exports = (modelName, schema) => {
