+++ lib/model/model-stub.js
@@ -1,7 +1,7 @@
 'use-strict';
 
-const modelQueryFunctions = require('../model-query-functions');
-const Query = require('./query');
+const modelQueryFunctions = require('./model-query-functions');
+const Query = require('../query/query-stub');
 
 module.exports = (modelName, schema) => {
   class Model {
