+++ lib/query/chainable-query-functions.js
