+++ lib/query/query-mock.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const deepEqual = require('deep-equal');
-const chainableQueryFunctions = require('../chainable-query-functions');
+const chainableQueryFunctions = require('./chainable-query-functions');
 
 // Mongoose Query docs: http://mongoosejs.com/docs/api.html#query-js
 // Mongoose Query code: https://github.com/Automattic/mongoose/blob/master/lib/query.js
