+++ lib/query/query-stub.test.js
@@ -2,10 +2,10 @@
 
 const chai = require('chai');
 const expect = chai.expect;
-const Query = require('./query');
-const chainableQueryFunctions = require('../../lib/chainable-query-functions');
+const Query = require('./query-stub');
+const chainableQueryFunctions = require('./chainable-query-functions');
 
-describe('Query Stub - Unit Tests', () => {
+describe('QueryStub', () => {
 
   // Automatically test that functions that are supposed to return queries all do the same thing
   describe('query return functions', () => {
